+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.0.0 IN PROGRESS
 
 * *BREAKING* Use `convertToSlipData` and supporting functions from `stripes-util`. Refs UIREQMED-92.
+* Remove validation from Request type field. Refs UIREQMED-97.
 
 ## [3.0.0] (https://github.com/folio-org/ui-requests-mediated/tree/v3.0.0) (2025-03-14)
 [Full Changelog](https://github.com/folio-org/ui-requests-mediated/compare/v2.0.2...v3.0.0)